+++ src/Amadeus/Client/RequestOptions/FareTLAGetFareRulesOptions.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Amadeus\Client\RequestOptions;
+
+class FareTLAGetFareRulesOptions extends Base
+{
+    /**
+     * @var string
+     */
+    public $airlineCode;
+
+    /**
+     * @var string
+     */
+    public $tariffClassId;
+}